requirements.txt
langgraph==0.1.19
langchain==0.2.6
langchain-groq==0.1.6
langchain-community==0.2.6
groq==0.9.0
chromadb==0.5.3
sentence-transformers==3.0.1
youtube-transcript-api==0.6.2
subliminal==2.2.0
babelfish==0.6.1
guessit==3.8.0
youtube-search-python==1.6.6
requests==2.32.3
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.134.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic
tenacity==8.3.0
python-dateutil==2.9.0
